+++ src/ir/mod.rs
@@ -7,6 +7,7 @@ pub mod function_emit;
 pub mod generate_loop;
 pub mod param;
 pub mod soft_union;
+pub mod task_emit;
 pub mod types;
 pub mod validate;